+++ service/pom.xml
@@ -6,7 +6,7 @@
     <packaging>jar</packaging>
 
     <parent>
-        <groupId>com.app-dev-project</groupId>
+        <groupId>com.wilma</groupId>
         <artifactId>wilma</artifactId>
         <version>0.0.1</version>
     </parent>
@@ -17,7 +17,7 @@
 
     <dependencies>
         <dependency>
-            <groupId>com.app-dev-project</groupId>
+            <groupId>com.wilma</groupId>
             <artifactId>repository</artifactId>
             <version>0.0.1</version>
             <scope>compile</scope>
