+++ web/pom.xml
@@ -6,7 +6,7 @@
     <packaging>jar</packaging>
 
     <parent>
-        <groupId>com.app-dev-project</groupId>
+        <groupId>com.wilma</groupId>
         <artifactId>wilma</artifactId>
         <version>0.0.1</version>
     </parent>
@@ -18,17 +18,17 @@
     <dependencies>
         <!-- Dependant project modules -->
         <dependency>
-            <groupId>com.app-dev-project</groupId>
+            <groupId>com.wilma</groupId>
             <artifactId>repository</artifactId>
             <version>0.0.1</version>
         </dependency>
         <dependency>
-            <groupId>com.app-dev-project</groupId>
+            <groupId>com.wilma</groupId>
             <artifactId>entity</artifactId>
             <version>0.0.1</version>
         </dependency>
         <dependency>
-            <groupId>com.app-dev-project</groupId>
+            <groupId>com.wilma</groupId>
             <artifactId>service</artifactId>
             <version>0.0.1</version>
         </dependency>
