+++ web/src/main/java/com/wilma/web/controller/HomeController.java
@@ -1,4 +1,4 @@
-package com.app_dev_project.controller;
+package com.wilma.web.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
